Moved value objects into one file.
