+++ ftt/handlers/portfolio_steps/portfolio_update_step.py
@@ -3,7 +3,7 @@ from typing import Optional
 from result import Ok, Err, Result
 
 from ftt.handlers.handler.abstract_step import AbstractStep
-from ftt.storage.data_objects import ValueObjectInterface
+from ftt.storage.value_objects import ValueObjectInterface
 from ftt.storage.errors import PersistingError
 from ftt.storage.models import Portfolio
 from ftt.storage.repositories.portfolios_repository import PortfoliosRepository
