+++ ftt/handlers/portfolio_version_steps/portfolio_version_associate_empty_weights_step.py
@@ -1,7 +1,7 @@
 from result import Ok
 
 from ftt.handlers.handler.abstract_step import AbstractStep
-from ftt.storage.data_objects.security_dto import SecurityValueObject
+from ftt.storage.value_objects import SecurityValueObject
 from ftt.storage.models import PortfolioVersion
 from ftt.storage.repositories.securities_repository import SecuritiesRepository
 from ftt.storage.repositories.weights_repository import WeightsRepository
