+++ ftt/handlers/securities_steps/securities_info_download_step.py
@@ -7,7 +7,7 @@ import yfinance as yf
 from result import Err, Ok, Result
 
 from ftt.handlers.handler.abstract_step import AbstractStep
-from ftt.storage.data_objects.security_dto import SecurityValueObject
+from ftt.storage.value_objects import SecurityValueObject
 
 
 class SecuritiesInfoDownloadStep(AbstractStep):
