+++ ftt/handlers/security_prices_steps/security_prices_upsert_step.py
@@ -4,7 +4,7 @@ import pandas as pd
 from result import Ok, Result
 
 from ftt.handlers.handler.abstract_step import AbstractStep
-from ftt.storage.data_objects.portfolio_version_dto import PortfolioVersionValueObject
+from ftt.storage.value_objects import PortfolioVersionValueObject
 from ftt.storage.mappers.security_price_mapper import SecurityPriceMapper
 from ftt.storage.models import PortfolioVersion
 from ftt.storage.repositories.securities_repository import SecuritiesRepository
