--- ftt/storage/data_objects/portfolio_dto.py
@@ -1,8 +1,0 @@
-from dataclasses import dataclass
-
-from ftt.storage.data_objects import ValueObjectInterface
-
-
-@dataclass
-class PortfolioValueObject(ValueObjectInterface):
-    name: str
