--- ftt/storage/data_objects/portfolio_security_prices_range_dto.py
@@ -1,10 +1,0 @@
-from dataclasses import dataclass
-from datetime import datetime
-
-from ftt.storage.data_objects import ValueObjectInterface
-
-
-@dataclass
-class PortfolioSecurityPricesRangeValueObject(ValueObjectInterface):
-    prices: dict[str, list]
-    datetime_list: list[datetime]
