--- ftt/storage/data_objects/portfolio_version_dto.py
@@ -1,15 +1,0 @@
-from dataclasses import dataclass
-from datetime import datetime
-from decimal import Decimal
-from typing import Optional
-
-from ftt.storage.data_objects import ValueObjectInterface
-
-
-@dataclass
-class PortfolioVersionValueObject(ValueObjectInterface):
-    period_start: Optional[datetime] = None
-    period_end: Optional[datetime] = None
-    interval: Optional[str] = None
-    value: Optional[Decimal] = None
-    optimization_strategy_name: Optional[str] = None
