--- ftt/storage/data_objects/weight_dto.py
@@ -1,11 +1,0 @@
-from dataclasses import dataclass
-from typing import Optional
-
-from ftt.storage.data_objects import ValueObjectInterface
-
-
-@dataclass
-class WeightValueObject(ValueObjectInterface):
-    planned_position: Optional[float] = None
-    position: Optional[float] = None
-    amount: Optional[float] = None
