+++ tests/handlers/order_steps/test_order_update_step.py
@@ -10,7 +10,7 @@ class TestOrderUpdateStep:
         return OrderUpdateStep
 
     def test_process_updates_order(self, subject, order):
-        from ftt.storage.data_objects.order_dto import OrderValueObject
+        from ftt.storage.value_objects import OrderValueObject
 
         dto = OrderValueObject(
             status=Order.Status.PARTIAL,
