+++ tests/handlers/securities_steps/test_securities_info_download_step.py
@@ -4,7 +4,7 @@ from result import Ok, Err
 from ftt.handlers.securities_steps.securities_info_download_step import (
     SecuritiesInfoDownloadStep,
 )
-from ftt.storage.data_objects.security_dto import SecurityValueObject
+from ftt.storage.value_objects import SecurityValueObject
 
 
 class TestSecuritiesInfoDownloadStep:
