+++ tests/handlers/test_portfolio_version_updation_handler.py
@@ -3,7 +3,7 @@ import pytest
 from ftt.handlers.portfolio_version_updation_handler import (
     PortfolioVersionUpdationHandler,
 )
-from ftt.storage.data_objects.portfolio_version_dto import PortfolioVersionValueObject
+from ftt.storage.value_objects import PortfolioVersionValueObject
 
 
 class TestPortfolioVersionUpdateHandler:
